+++ src/languages/en.json
@@ -1,19 +1,19 @@
 {
-    "sidebar-context.addImage": "Add Image",
-    "sidebar-context.prototype": "Prototype Token",
-    "sidebar-context.displayChat": "Display in Chat",
-    "sidebar-context.updateChildren": "Update Child Tokens",
-    "sidebar-context.updateChildrenTitle": "Update all token data",
-    "sidebar-context.updateChildrenContent": "This will update all tokens created from this actor, to match the new prototype token data",
-    "sidebar-context.resetDoors": "Reset Doors",
-    "sidebar-context.resetFog": "Reset Fog",
-    "sidebar-context.showNavAll": "Show all scenes in navigation bar",
-    "sidebar-context.hideNavAll": "Hide all scenes in navigation bar",
-    "sidebar-context.requestPrompt": "Can someone roll a result from this random table?",
-    "sidebar-context.requestButton": "Let's Get Random!",
-    "sidebar-context.playerThanks": "Thanks {PLAYER}",  
-    "sidebar-context.menuMakeRoll": "Make Roll",
-    "sidebar-context.menuRequestRoll": "Request Named Roll",
-    "sidebar-context.menuRequestBlindRoll": "Request Blinded Roll",
-    "sidebar-context.menuRequestDescRoll": "Request Descriptive Roll"
+	"sidebar-context.addImage": "Add Image",
+	"sidebar-context.prototype": "Prototype Token",
+	"sidebar-context.displayChat": "Display in Chat",
+	"sidebar-context.updateChildren": "Update Child Tokens",
+	"sidebar-context.updateChildrenTitle": "Update all token data",
+	"sidebar-context.updateChildrenContent": "This will update all tokens created from this actor, to match the new prototype token data",
+	"sidebar-context.resetDoors": "Reset Doors",
+	"sidebar-context.resetFog": "Reset Fog",
+	"sidebar-context.showNavAll": "Show all scenes in navigation bar",
+	"sidebar-context.hideNavAll": "Hide all scenes in navigation bar",
+	"sidebar-context.requestPrompt": "Can someone roll a result from this random table?",
+	"sidebar-context.requestButton": "Let's Get Random!",
+	"sidebar-context.playerThanks": "Thanks {PLAYER}",
+	"sidebar-context.menuMakeRoll": "Make Roll",
+	"sidebar-context.menuRequestRoll": "Request Named Roll",
+	"sidebar-context.menuRequestBlindRoll": "Request Blinded Roll",
+	"sidebar-context.menuRequestDescRoll": "Request Descriptive Roll"
 }
