+++ src/main.ts
@@ -1,0 +1,64 @@
+/**
+ * This is your TypeScript entry file for Foundry VTT.
+ * Register custom settings, sheets, and constants using the Foundry API.
+ * Change this heading to be more descriptive to your module, or remove it.
+ * Author: [your name]
+ * Content License: [copyright and-or license] If using an existing system
+ * 					you may want to put a (link to a) license or copyright
+ * 					notice here (e.g. the OGL).
+ * Software License: [your license] Put your desired license here, which
+ * 					 determines how others may use and modify your module
+ */
+// Import JavaScript modules
+
+// Import TypeScript modules
+import CONSTANTS from "./scripts/constants";
+import { initHooks } from "./scripts/sidebar-context";
+import { registerSettings } from "./scripts/settings";
+
+/* ------------------------------------ */
+/* Initialize module					*/
+/* ------------------------------------ */
+Hooks.once("init", async () => {
+	console.log(`${CONSTANTS.MODULE_NAME} | Initializing ${CONSTANTS.MODULE_NAME}`);
+
+	// Register custom module settings
+	registerSettings();
+
+	initHooks();
+	// Assign custom classes and constants here
+
+	// Register custom module settings
+	//registerSettings();
+	//fetchParams();
+
+	// Preload Handlebars templates
+	// await preloadTemplates();
+	// Register custom sheets (if any)
+});
+
+/* ------------------------------------ */
+/* Setup module							*/
+/* ------------------------------------ */
+Hooks.once("setup", function () {
+	// Do anything after initialization but before ready
+	//setupModules();
+	// setupHooks();
+	//registerSettings();
+});
+
+/* ------------------------------------ */
+/* When ready							*/
+/* ------------------------------------ */
+Hooks.once("ready", () => {
+	// Do anything once the module is ready
+	// readyHooks();
+});
+
+/* ------------------------------------ */
+/* Other Hooks							*/
+/* ------------------------------------ */
+
+Hooks.once("devModeReady", ({ registerPackageDebugFlag }) => {
+	registerPackageDebugFlag(CONSTANTS.MODULE_NAME);
+});
