+++ src/module.json
@@ -1,14 +1,14 @@
 {
-  "id": "sidebar-context",
-  "title": "Sidebar Context",
-  "description": "Adds several right click context options to documents",
-  "version": "This is auto replaced",
+	"id": "sidebar-context",
+	"title": "Sidebar Context",
+	"description": "Adds several right click context options to documents",
+	"version": "0.1.1",
 	"authors": [
-    {
-      "name": "Kandashi",
-      "url": "https://github.com/kandashi",
-      "discord": "Kandashi (He/Him)#6698"
-    },
+		{
+			"name": "Kandashi",
+			"url": "https://github.com/kandashi",
+			"discord": "Kandashi (He/Him)#6698"
+		},
 		{
 			"name": "p4535992",
 			"url": "https://github.com/p4535992",
@@ -59,25 +59,21 @@
 	"packs": [],
 	"systems": [],
 	"scripts": [],
-  "esmodules": [
-    "/scripts/sidebarContext.js"
-  ],
-  "styles": [
-    "/styles/module.css"
-  ],
+	"esmodules": ["/scripts/main.js"],
+	"styles": ["/styles/sidebar-context.css"],
 	"compatibility": {
 		"minimum": 10,
 		"verified": 10,
 		"maximum": 10
 	},
 	"manifestPlusVersion": "1.2.1",
-  "url": "This is auto replaced",
-  "manifest": "This is auto replaced",
-  "download": "This is auto replaced",
-  "readme": "https://github.com/kandashi/sidebar-context/blob/master/README.md",
-  "changelog": "https://github.com/kandashi/sidebar-context/blob/master/changelog.md",
-  "bugs": "https://github.com/kandashi/sidebar-context/issues",
-  "allowBugReporter": true,
+	"url": "0.1.1",
+	"manifest": "https://github.com/p4535992/foundryvtt-sidebar-context/releases/latest/download/module.json",
+	"download": "https://github.com/p4535992/foundryvtt-sidebar-context/releases/latest/download/module.zip",
+	"readme": "https://github.com/p4535992/foundryvtt-sidebar-context/blob/master/README.md",
+	"changelog": "https://github.com/p4535992/foundryvtt-sidebar-context/blob/master/changelog.md",
+	"bugs": "https://github.com/p4535992/foundryvtt-sidebar-context/issues",
+	"allowBugReporter": true,
 	"relationships": {
 		"systems": [],
 		"requires": [
