+++ src/scripts/constants.ts
@@ -1,0 +1,10 @@
+const CONSTANTS = {
+	MODULE_NAME: "sidebar-context",
+	PATH: `modules/sidebar-context/`,
+	TEMPLATE_PATH: `/modules/sidebar-context/templates`,
+	WHISPER_FN: "cheekyWhisper",
+};
+
+CONSTANTS.PATH = `modules/${CONSTANTS.MODULE_NAME}/`;
+
+export default CONSTANTS;
