+++ src/scripts/settings.ts
@@ -1,0 +1,10 @@
+export const registerSettings = function (): void {
+	// game.settings.register(CONSTANTS.MODULE_NAME, "applyOnATEItem", {
+	// 	name: i18n(`${CONSTANTS.MODULE_NAME}.setting.applyOnATEItem.name`),
+	// 	hint: i18n(`${CONSTANTS.MODULE_NAME}.setting.applyOnATEItem.hint`),
+	// 	scope: "world",
+	// 	config: true,
+	// 	type: Boolean,
+	// 	default: true,
+	// });
+};
